Fix review comments - Add objects to FW components - Clean test file to contain only test cases - Conftest contains fixtures
